refactor: replace ParentChain with NoteHeader for all notes
- Create new NoteHeader component that shows for all notes
- Extract header bar logic into reusable component
- Update SearchView to use NoteHeader instead of conditional ParentChain
- Remove old ParentChain component
- Keep DRY principles and prepare for future icon/button additions
